improved docker scripts

+++ docker/build_compass.sh
@@ -7,7 +7,10 @@ echo "os: '$OS_NAME'"
 
 export QT_SELECT=5
 
-rm -rf /app/workspace/compass/build_$OS_NAME # needed since binary becomes too big
+if [[ "$1" == "--clean" ]]; then
+    rm -rf /app/workspace/compass/build_$OS_NAME
+fi
+
 mkdir -p /app/workspace/compass/build_$OS_NAME
 cd /app/workspace/compass/build_$OS_NAME
 cmake -DCMAKE_INSTALL_PREFIX=/usr -DCMAKE_BUILD_TYPE=RelWithDebInfo .. # -DCMAKE_PREFIX_PATH=$GL_PATH $CMAKE_OPTS
